requirements.txt
# Road Detection Pipeline — Python Dependencies

# Core
torch>=2.0.0
torchvision>=0.15.0
rasterio>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
Pillow>=9.0.0

# TorchGeo + Segmentation
torchgeo>=0.6.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
